VC-1 interlaced B-frames support by updating VC-1 related files closer to ffmpeg master.
