+++ lib/ffmpeg/libavcodec/vc1_parser.c
@@ -57,16 +57,16 @@ static void vc1_extract_headers(AVCodecParserContext *s, AVCodecContext *avctx,
         if(size <= 0) continue;
         switch(AV_RB32(start)){
         case VC1_CODE_SEQHDR:
-            vc1_decode_sequence_header(avctx, &vpc->v, &gb);
+            ff_vc1_decode_sequence_header(avctx, &vpc->v, &gb);
             break;
         case VC1_CODE_ENTRYPOINT:
-            vc1_decode_entry_point(avctx, &vpc->v, &gb);
+            ff_vc1_decode_entry_point(avctx, &vpc->v, &gb);
             break;
         case VC1_CODE_FRAME:
             if(vpc->v.profile < PROFILE_ADVANCED)
-                vc1_parse_frame_header    (&vpc->v, &gb);
+                ff_vc1_parse_frame_header    (&vpc->v, &gb);
             else
-                vc1_parse_frame_header_adv(&vpc->v, &gb);
+                ff_vc1_parse_frame_header_adv(&vpc->v, &gb);
 
             /* keep AV_PICTURE_TYPE_BI internal to VC1 */
             if (vpc->v.s.pict_type == AV_PICTURE_TYPE_BI)
@@ -196,6 +196,6 @@ AVCodecParser ff_vc1_parser = {
     .priv_data_size = sizeof(VC1ParseContext),
     .parser_init    = vc1_parse_init,
     .parser_parse   = vc1_parse,
-    .parser_close   = ff_parse1_close,
+    .parser_close   = ff_parse_close,
     .split          = vc1_split,
 };
